# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SbCommunicationTest",
        "description": "Back End Application For Parcel delivery system",
        "contact": {
            "email": "danieldada123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/user/register": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create User",
                "description": "Create user with name  ",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Login User",
                "description": "login  ",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "description": "password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/create/parcel": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create Parcel details",
                "description": "--  ",
                "operationId": "CreateParcel",
                "parameters": [
                    {
                        "name": "weight",
                        "in": "path",
                        "description": "Weight of the Parcel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_address",
                        "in": "path",
                        "description": "Address of the sender",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to_address",
                        "in": "path",
                        "description": "Address of Receiver",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/updateParcelDestination/{id}'": {
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update Parcel destination",
                "description": "--  ",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "to_address",
                        "in": "path",
                        "description": "Address of the receiver",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/cancelParcelDelivery/{id}'": {
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Cancel Parcel ",
                "description": "--  ",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the parcel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/updateEnrouteDelivery/{id}'": {
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update Parcel status",
                "description": "--  ",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the parcel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/updateCompleteDelivery/{id}'": {
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Complete Parcel status",
                "description": "--  ",
                "operationId": "Complete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the parcel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/user/getParcelDetails/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Fetch Parcel detail",
                "description": "--  ",
                "operationId": "Fetch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the parcel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        }
    }
}